requirements.txt
playwright==1.38.0
opencv-python==4.8.0.76
numpy==1.24.3
Pillow==10.0.0
aiohttp==3.8.5
pygetwindow==0.0.9
mss==9.0.1 